# RFC: perf(base): do not unconditionally process cmdline with each `getcmdline` call