--- a/modules.d/35network-legacy/parse-ip-opts.sh
+++ b/modules.d/35network-legacy/parse-ip-opts.sh
@@ -43,6 +43,7 @@ BOOTDEV=$(getarg bootdev=)
 if [ -n "$NEEDBOOTDEV" ] && getargbool 1 rd.neednet; then
     #[ -z "$BOOTDEV" ] && warn "Please supply bootdev argument for multiple ip= lines"
     echo "rd.neednet=1" > /etc/cmdline.d/dracut-neednet.conf
+    setcmdline
     info "Multiple ip= arguments: assuming rd.neednet=1"
 else
     unset NEEDBOOTDEV
@@ -131,6 +132,7 @@ for p in $(getargs ip=); do
             "$(expr substr "$dev" 11 2)" \
             "$(expr substr "$dev" 13 2)" \
             >> /etc/cmdline.d/80-enx.conf
+        setcmdline
     fi
 done