@aafeijoo-suse 2023-04-25
> > This patch adds a new function `setcmdline`, which parses command line options, exports the `CMDLINE` variable (required by `dracut-util`), and persists its content in a temporary file, so that different processes can access it (e.g. systemd generators).
> 
> Just a thought - can we avoid creating some of the /etc/cmdline.d files ?
> 
> As an example change
> 
> ```
> echo rd.neednet >> /etc/cmdline.d/connman.conf
> setcmdline
> ```
> 
> to
> 
> ```
> echo rd.neednet | setcmdline
> ```
> 

We could, but I don't know if it's a good idea. If we remove these conf files, it would be more difficult to debug from the emergency shell which module adds command line parameters.

> > With this change, if a hook script adds new content in /etc/cmdline.d at boot, it must explicitly call setcmdline to update the content of the CMDLINE variable.
> 
> Would this break out-of-tree dracut modules ?

AFAIK the only documented use of /etc/cmdline.d to inject custom command line options is during the build process, not dynamically at boot. If an external dracut module is doing this (e.g. [openSUSE kdump](https://github.com/openSUSE/kdump/blob/eed2330577c0e1a4b392e0cc4c6227ca93be9d5b/dracut/kdump-boot.sh#L19)), then yes, it will need to adapt its code. But I think this should not be an obstacle to fix this inefficient code.