@FGrose 2023-05-17
This will also clean up debug xtrace logs.